requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
langchain-groq        # ← ADD THIS
langgraph             # ← ADD THIS
python-dotenv
pydantic
